refactor: split responder god-object into cohesive files; dedup helpers
Split responder/responder.go (1087 lines) into same-package files by
responsibility, with zero public-API or behavior changes:
  - responder.go      struct, New, Close, IP + dedup helpers (63 LOC core)
  - lifecycle.go      Register, Unregister, GetService, UpdateService
  - query_handler.go  runQueryHandler, validateSourceAddress, handleQuery
  - testhooks.go      test-only observation/injection hooks (clearly segregated)

Deduplication (single source of truth):
  - buildServiceInfo()        records.ServiceInfo built once (was 3x inline)
  - to/fromInternalService()  public<->internal Service conversion (was 4x inline)
  - protocol.MulticastGroupIPv4() replaces 2 inline 224.0.0.251:5353 literals

Hygiene:
  - remove stale //lint:ignore U1000 on getIPv4ForInterface (verified used)
  - replace 'if err != nil { _ = err }' noop with best-effort send + comment
  - document UpdateService best-effort announcement semantics

Verification: full suite + -race green; two independent reviews confirm
behavior-preserving; live two-process wire test matches main baseline.
